Completed box generator assignment achieved MVP and added new features:
I have color and height/width as inputs and then use those inputs to generate
a box with the corresponding background color and dimensions. I added data
persistence in local storage (to persist through refreshes) and ref hooks
in order to clear out input data and refocus on color input.